% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAutoxgboostMC.R
\name{shinyAutoxgboostMC}
\alias{shinyAutoxgboostMC}
\title{Run shinyAutoxgboostMC}
\usage{
shinyAutoxgboostMC(...)
}
\arguments{
\item{...}{[\code{any}]\cr
Additional arguments passed to shiny's
\code{runApp()} function.}
}
\description{
Run a local instance of shinyAutoxgboostMC
}
\examples{
\dontrun{
  shinyAutoxgboostMC()
}
}
